+++ util/login.py
@@ -1,0 +1,26 @@
+from functools import wraps
+from flask import redirect
+from flask_jwt_extended import get_jwt_identity, verify_jwt_in_request
+from models.user import User
+
+
+def login_required(fn):
+    """ Wraps a function and injects a login check before each call. Redirects
+    to the login page if the current user is not logged in.
+    :param fn: the function to wrap.
+    """
+
+    # pylint: disable=duplicate-code
+    @wraps(fn)
+    # pylint: disable=duplicate-code
+    def wrapper(*args, **kwargs):
+        # pylint: disable=duplicate-code
+        verify_jwt_in_request()
+        # pylint: disable=duplicate-code
+        user_id = get_jwt_identity()
+        # pylint: disable=duplicate-code
+        target_user = User.query.filter_by(id=user_id).first()
+        if target_user is None:
+            return redirect('/admin/login', code=403)
+        return fn(*args, **kwargs)
+    return wrapper
